Update simple_budget.rb
Fixed a bug that caused the function "member_register_loop" to run twice no matter the specified value.
Fixed an issue where the first user to enter the information was the only user prompted for information.

Instead of adding each member's information to an array inside of the "member_register_loop" function, the code was trying to add it in when the function was called, which caused the above bugs. The "member_register_loop" function now works as intended when called.